+++ src/client/src/pages/overview/layout/mainSideNav.tsx
@@ -17,7 +17,7 @@ import { ImportErrorDialog } from "../sidePanelContent/importer/importErrorDialo
 export interface MainSideNavProps {
   toggleDrawer: (open: boolean) => void;
   drawerOpen: boolean;
-  setWorkgroupId: React.Dispatch<React.SetStateAction<string>>;
+  setWorkgroupId: React.Dispatch<React.SetStateAction<number | null>>;
   setEnabledWorkgroups: React.Dispatch<React.SetStateAction<Workgroup[]>>;
   setSideDrawerContent: React.Dispatch<React.SetStateAction<DrawerContentTypes>>;
   sideDrawerContent: DrawerContentTypes;
@@ -48,7 +48,7 @@ const MainSideNav = ({
   useEffect(() => {
     const wgs = user.data.workgroups.filter(w => w.disabled === null && w.roles.includes("EDIT"));
     setEnabledWorkgroups(wgs);
-    setWorkgroupId(wgs.length > 0 ? wgs[0].id : "");
+    setWorkgroupId(wgs.length > 0 ? wgs[0].id : null);
   }, [setEnabledWorkgroups, setWorkgroupId, user.data.workgroups]);
 
   const handleToggleFilter = () => {